fix: _needs_sudo() walks up tree for non-existent paths; fix 99-cleanup typos

+++ install/post-install/99-cleanup.sh
@@ -3,9 +3,9 @@
 # We have nothing to connect / listen to in a vm.
 if [[ "$CHASSIS_TYPE" == "vm" ]]; then
   sudo rm -rf /etc/udev/rules.d/*
-elelif [[ "$CHASSIS_TYPE" == "desktop" ]]; then
+elif [[ "$CHASSIS_TYPE" == "desktop" ]]; then
   sudo rm -f /etc/udev/rules.d/10-battery-alert.rules
-  surp rmf -f /etc/udev/rules.d/10-power-profile.rules
+  sudo rm -f /etc/udev/rules.d/10-power-profile.rules
 fi
 
 yay -Rnsc yay-debug --noconfirm &>/dev/null || true
